refactor(CHATROOM): CHAT MESSAGE - 각 메시지별 구성요소 분리
