+++ src/chat/talk/chatRoom/chatMessage.tsx
@@ -2,6 +2,7 @@ import React, { useContext } from 'react';
 
 import SystemMessage from './systemMessage';
 import PartnerMessage from './partnerMessage';
+import UserMessage from './userMessage';
 
 import { ModalContext } from '../../store/modalContext';
 
@@ -37,33 +38,6 @@ const ChatMessage: React.FC<ChatMessageProps> = (props) => {
         <>
             {
                 messages.map((current, idx) => {
-                    const messageStyle: React.CSSProperties = {
-                        maxWidth: "58%",
-                        padding: "8px",
-                        margin: "8px",
-                        fontSize: "1rem",
-                        color: "#1b1b1b",
-                        wordBreak: "break-word",
-                        background: "#fff",
-                        borderRadius: "6px",
-                    };
-                    const boxStyle: React.CSSProperties = {
-                        display: "flex",
-                        flex: "1",
-                        flexFlow: "column",
-                        width: "100%",
-                        margin: 0,
-                    };
-                    if (current.idx === userId) {
-                        messageStyle.background = "tomato";
-                        boxStyle.marginLeft = "auto";
-                        boxStyle.flexFlow = "row-reverse";
-                    }
-                    else if (current.idx !== '#system') {
-                        boxStyle.marginRight = "auto";
-                        boxStyle.flexFlow = "row";
-                    }
-
                     return (
                         <article
                             key={idx}
@@ -73,67 +47,29 @@ const ChatMessage: React.FC<ChatMessageProps> = (props) => {
                             }}
                         >
                             {
-                                current.idx !== '#system' && current.idx !== userId &&
-                                <span
-                                    style={{
-                                        paddingLeft: "8px",
-                                        fontSize: "0.8rem",
-                                    }}
-                                >
-                                    {current.idx}
-                                </span>
+                                current.idx === '#system' &&
+                                <SystemMessage
+                                    message={current.message}
+                                />
                             }
-                            <p style={boxStyle}>
-                                {
-                                    current.idx === '#system' &&
-                                    <SystemMessage
-                                        message={current.message}
-                                    />
-                                }
-                                {
-                                    current.idx !== '#system' && current.idx !== userId &&
-                                    <PartnerMessage
-                                        message={current.message}
-                                        openImageModal={openImageModal}
-                                    />
+                            {
+                                current.idx !== '#system' && current.idx !== userId &&
+                                <PartnerMessage
+                                    idx={current.idx}
+                                    message={current.message}
+                                    openImageModal={openImageModal}
+                                    time={current.time}
+                                />
 
-                                }
-                                {
-                                    current.idx !== '#system' && current.idx === userId &&
-                                    (current.message.slice(0, 5) !== '@$IMG' ?
-                                        <span
-                                            style={{
-                                                maxWidth: "58%",
-                                                padding: "8px",
-                                                margin: "8px",
-                                                fontSize: "1rem",
-                                                color: "#1b1b1b",
-                                                wordBreak: "break-word",
-                                                background: "tomato",
-                                                borderRadius: "6px",
-                                            }}
-                                        >
-                                            {current.message}
-                                        </span> :
-                                        <img
-                                            src={current.message.slice(5)}
-                                            style={messageStyle}
-                                            onClick={() => openImageModal(current.message.slice(5))}
-                                        />)
-                                }
-                                {
-                                    current.idx !== '#system' &&
-                                    < time
-                                        style={{
-                                            display: "flex",
-                                            alignItems: "flex-end",
-                                            marginBottom: "8px",
-                                            fontSize: "0.7rem",
-                                        }}
-                                    >{current.time}</time>
-                                }
-                            </p>
-
+                            }
+                            {
+                                current.idx !== '#system' && current.idx === userId &&
+                                <UserMessage
+                                    message={current.message}
+                                    openImageModal={openImageModal}
+                                    time={current.time}
+                                />
+                            }
                         </article>
                     );
                 })
