+++ src/chat/talk/chatRoom/userMessage.test.tsx
@@ -9,21 +9,25 @@ describe('유저 메시지', () => {
             <UserMessage
                 message='test'
                 openImageModal={() => { }}
+                time='오후 12:23'
             />
         );
 
         expect(screen.getByTestId('user')).toBeInTheDocument();
         expect(screen.getByText('test')).toBeInTheDocument();
+        expect(screen.getByText('오후 12:23')).toBeInTheDocument();
     });
     test('render img', () => {
         render(
             <UserMessage
                 message='@$IMG img'
                 openImageModal={() => { }}
+                time='오후 12:23'
             />
         );
 
         expect(screen.getByTestId('user-image')).toBeInTheDocument();
         expect(screen.getByAltText('전송된 이미지')).toBeInTheDocument();
+        expect(screen.getByText('오후 12:23')).toBeInTheDocument();
     });
 });
